% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_1.R
\name{DGM_1}
\alias{DGM_1}
\title{Generate longitudinal dataset}
\usage{
DGM_1(n = 1000, m = 3, b = NULL, beta, sigma = 1)
}
\arguments{
\item{n}{sample size; default is 1000}

\item{m}{number of visits per subject; default is 3}

\item{b}{\code{n}-dimensional vector of random intercepts; if NULL, random intercept is generated from N(0, 1)}

\item{beta}{vector of regression coefficients on \code{z_y}, used for generation of \code{y}}

\item{sigma}{standard deviation of the random error, epsilon; default is 1}
}
\description{
Generate data set for \code{n} clusters each of size \code{m}
with a response \code{y}, subject index \code{id}, as well as
continuous and time-dependent covariates \code{z_y}
}
