% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_score_vector.R
\name{eff_score_vec}
\alias{eff_score_vec}
\title{Calculate the efficient score function for theta}
\usage{
eff_score_vec(data, response)
}
\arguments{
\item{data}{dataframe for one cluster of observations}

\item{response}{character denoting the column name for the model response}
}
\description{
Take in in dataframe 'data', which corresponds to data from a
single cluster, then return the efficient score function thot
beta corresponding to that cluster.
Intended for use with geex() function
}
