% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnereate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate longitudinal dataset}
\usage{
generate_data(n = 1000, m = 3, beta, b = NULL, sigma = 1, p.X = 1, p.A = 1)
}
\arguments{
\item{n}{sample size; default is 1000}

\item{m}{number of visits per subject; default is 3}

\item{beta}{3-dimensional vector of regression coefficients}

\item{b}{\code{n}-dimensional vector of random intercepts; if NULL, random intercept is generated from N(0, 1)}

\item{sigma}{standard deviation of the random error, epsilon; default is 1}

\item{p.X}{number of time-variant, continuous covariates}

\item{p.A}{number of time-invariant, binary covariates}
}
\description{
Generate data set for \code{n} clusters each of size \code{m}
with a response Y, 1 binary, time-invariant covariate X1,
1 continuous, time-variant covariate X2, and subject index id
}
