% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_score_vector.R
\name{eff_score_vec}
\alias{eff_score_vec}
\title{This function calculates the efficient score vector proposed by
Grosser et al. (2022)}
\usage{
eff_score_vec(data, response, X.names, Z.names = NULL, cens = NULL)
}
\arguments{
\item{data}{dataframe for one cluster of observations}

\item{response}{character denoting the column name for the model response}

\item{X.names}{string of characters denoting the columns for the model covariates that are associated with fixed effects}

\item{Z.names}{string of characters denoting the columns for the model covariates that are associated with random effects.
Default value is NULL, in which case the random effects design matrix is assumed to be a column of 1's}

\item{cens}{character denoting the column name for the censoring indicator, which is 1 if not censored and 0 if censored.
Default value is NULL, in which case the cluster is considered not censored.}
}
\description{
Estimates the parameters of a linear mixed model when we do
not want to specify a distribution for the random effects and
when one covariate is (possibly) subject to random right-censoring
Takes in dataframe 'data', which corresponds to data from a
single cluster, then returns the value of efficient score vector for
theta = (beta, sigma^2) corresponding to that cluster.
Intended for use with m_estimate() function
At the moment, this function only allows for random intercept.
}
