% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_score_vector.R
\name{eff_score_vec}
\alias{eff_score_vec}
\title{Calculate the efficient score function for theta}
\usage{
eff_score_vec(data, response, X.names, cens = NULL)
}
\arguments{
\item{data}{dataframe for one cluster of observations}

\item{response}{character denoting the column name for the model response}

\item{X.names}{string of characters denoting the columns for the model covariates that are associated with fixed effects}

\item{cens}{character denoting the column name for the censoring indicator; 1 is not censored and 0 is censored}
}
\description{
Take in in dataframe 'data', which corresponds to data from a
single cluster, then return the efficient score function for
theta = (beta, alpha, sigma^2) corresponding to that cluster.
Intended for use with m_estimate() function
Right now, intended for use in linear mixed model when we do
not want to specify a distribution for the random effects.
Also, at the moment, this function only allows for random
intercept.
}
